requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0

# Image processing
opencv-python>=4.5.0
pillow>=8.0.0

# 3D processing and visualization
open3d>=0.13.0
laspy>=2.0.0

# Point cloud specific
# pdal>=2.4.0 устанавливать через conda: conda install -c conda-forge pdal

# Configuration and utilities
pyyaml>=5.4.0
python-dotenv>=0.19.0

# CLI
click>=8.0.0

# Visualization
matplotlib>=3.3.0

# Development (optional)
pytest>=6.0.0
pytest-cov>=2.12.0